requirements.txt
# Web API and server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
werkzeug==2.2.3
websockets

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Audio processing
pydub==0.25.1

# External APIs
assemblyai>=0.6.3
pinecone>=6.0.2
requests>=2.31.0

# Storage
boto3==1.29.3

# ML and data processing
numpy==1.26.2
torch==2.1.1